Fixed data bleeding
Found issue if a connection could not be made, the results from the previous query would be placed into the current file. I now clear the output results each time so this will not happen anymore.

I also now include the dbid for the enterprise feature query. This way we can link this back to the DBInfo file.