Streamlined input process and allow for earlier versions of SQL to work
Script will now work properly with earlier versions of SQL.

Script will also now get the data and log file paths from the backup itself. It will use those paths to do the restore on the secondary replica servers. This is valid in AG setups because AGs require that all things be setup the same on replica servers.