Add files via upload
This script builds on the previous Storage Policy to SLA Protection by allowing you the user to select what Storage Policy to associate with a SLA. Instead of needing to know what SLAs and Storage Policies are in your environment, it will query the vCenter and Rubrik machine for you.  In addition to this, I added the capability to capture two different credentials and allowed you to specify the vCenter and Rubrik brik name realtime. This script will continue to grow as I think up of new capabilities to add to it. At the end the script gives you a short summary of what VMs are still unprotected.